+++ packages/crypto/src/pbkdf2.ts
@@ -1,39 +1,41 @@
 import { CodeError } from '@libp2p/interface'
-// @ts-expect-error types are missing
-import forgePbkdf2 from 'node-forge/lib/pbkdf2.js'
-// @ts-expect-error types are missing
-import forgeUtil from 'node-forge/lib/util.js'
+import { pbkdf2 as pbkdf2Sync } from '@noble/hashes/pbkdf2'
+import { sha1 } from '@noble/hashes/sha1'
+import { sha256 } from '@noble/hashes/sha256'
+import { sha512 } from '@noble/hashes/sha512'
+import { base64 } from 'multiformats/bases/base64'
 
 /**
- * Maps an IPFS hash name to its node-forge equivalent.
+ * Maps an IPFS hash name to its @noble/hashes equivalent.
  *
  * See https://github.com/multiformats/multihash/blob/master/hashtable.csv
  *
  * @private
  */
 const hashName = {
-  sha1: 'sha1',
-  'sha2-256': 'sha256',
-  'sha2-512': 'sha512'
+  sha1,
+  'sha2-256': sha256,
+  'sha2-512': sha512
 }
 
 /**
  * Computes the Password-Based Key Derivation Function 2.
  */
-export default function pbkdf2 (password: string, salt: string, iterations: number, keySize: number, hash: string): string {
+export default function pbkdf2 (password: string, salt: string | Uint8Array, iterations: number, keySize: number, hash: string): string {
   if (hash !== 'sha1' && hash !== 'sha2-256' && hash !== 'sha2-512') {
     const types = Object.keys(hashName).join(' / ')
     throw new CodeError(`Hash '${hash}' is unknown or not supported. Must be ${types}`, 'ERR_UNSUPPORTED_HASH_TYPE')
   }
 
   const hasher = hashName[hash]
-  const dek = forgePbkdf2(
+  const dek = pbkdf2Sync(
+    hasher,
     password,
-    salt,
-    iterations,
-    keySize,
-    hasher
+    salt, {
+      c: iterations,
+      dkLen: keySize
+    }
   )
 
-  return forgeUtil.encode64(dek, null)
+  return base64.encode(dek).substring(1)
 }