+++ packages/protocol-identify/CHANGELOG.md
@@ -60,6 +60,14 @@
   * devDependencies
     * @libp2p/peer-id-factory bumped from ^4.0.3 to ^4.0.4
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/peer-record bumped from ^7.0.6 to ^7.0.7
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^5.1.3 to ^5.2.0
+
 ## [1.0.11](https://github.com/libp2p/js-libp2p/compare/identify-v1.0.10...identify-v1.0.11) (2024-01-16)